build(apps-cli): refactor nest command
feat: add cli app
