+++ apps/cli/src/environments/environment.ts
@@ -1,0 +1,3 @@
+export const environment = {
+  production: false,
+}
