+++ apps/cli/src/main.ts
@@ -1,0 +1,1 @@
+console.log('Hello World!')
