+++ apps/gqlgen/src/e2e/index.ts
@@ -1,0 +1,2 @@
+export * from './e2e.module'
+export * from './e2e.service'
