+++ apps/gqlgen/src/graphql-codegen/index.ts
@@ -1,0 +1,2 @@
+export * from './graphql-codegen.module'
+export * from './graphql-codegen.service'
