+++ apps/gqlgen/src/server/index.ts
@@ -1,0 +1,2 @@
+export * from './server.module'
+export * from './server.service'
