+++ apps/gqlgen/src/server/server.module.ts
@@ -1,4 +1,9 @@
-import { GraphqlServerConfig, GraphqlServerTokens } from '@codelab/backend'
+import {
+  GraphqlServerConfig,
+  GraphqlServerTokens,
+  serverConfig,
+  ServerTokens,
+} from '@codelab/backend'
 import { DynamicModule, Module } from '@nestjs/common'
 import { ConfigFactory } from '@nestjs/config'
 import { ServerService } from './server.service'
@@ -12,6 +17,10 @@ export class ServerModule {
         {
           provide: GraphqlServerTokens.GraphqlServerConfig,
           useValue: config(),
+        },
+        {
+          provide: ServerTokens.ServerConfig,
+          useValue: serverConfig(),
         },
       ],
       exports: [ServerService],
