+++ libs/backend/src/common/environment/environment.ts
@@ -1,4 +1,8 @@
 export enum Environment {
+  // Development, regular serve
   Dev = 'Dev',
-  Test = 'Test',
+  // This is the CI pre-push for local flow
+  Local = 'Local',
+  // This is the CircleCI flow
+  Ci = 'Ci',
 }
