+++ package.json
@@ -15,7 +15,8 @@
     "postinstall": "husky install",
     "nx-env": "env-cmd -f .env.test nx",
     "gqlgen:ci": "cross-env NODE_ENV=test node -r dotenv-flow/config dist/apps/gqlgen/main.js",
-    "gqlgen": "nx build gqlgen && node -r dotenv-flow/config dist/apps/gqlgen/main.js",
+    "gqlgen": "nx build gqlgen --skip-nx-cache && env-cmd -f .env node dist/apps/gqlgen/main.js",
+    "console": "env-cmd -f .env node dist/apps/gqlgen/main.js",
     "tunnel": "ngrok http -region=us -hostname=a8069b052300.ngrok.io http://127.0.0.1:3000",
     "docker:up": "libs/tools/scripts/src/docker/up.sh",
     "nx": "nx",
