+++ scripts/nx.sh
@@ -1,6 +1,5 @@
 #!/usr/bin/env bash
 
-
 set -x
 
 ARGS="$*"
